# server/python_server/src/http_server/models/api_response.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, validator  # noqa: F401


class ApiResponse(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    ApiResponse - a model defined in OpenAPI

        code: The code of this ApiResponse [Optional].
        message: The message of this ApiResponse [Optional].
        sub_message: The sub_message of this ApiResponse [Optional].
        data: The data of this ApiResponse [Optional].
        items: The items of this ApiResponse [Optional].
    """

    code: Optional[int] = None
    message: Optional[str] = None
    sub_message: Optional[str] = None
    data: Optional[Dict[str, Any]] = None
    items: Optional[List[object]] = None

ApiResponse.update_forward_refs()